Bugfix - Correct calculation of distances
Since I took the computations from the monte carlo implementation of the
harmonic potential there were some unusual elements to the calculation
of the position, I believe it was taking into account the move that was
being attempted. Since I am interested with the force at the current
point in time for moelcular dynamics, this required a somewhat different
approach. This additionally significantly simplifies the calculation
of the distance between the two points.