Correct the imports of and inheritance of modules
I was initially not sure how this plugin, once installed would interact
with hoomd, particularly in terms of how everything is imported. This
performs the correct inheritance, from `_force` instead of `_compute`
and uses relative imports rather than absolute imports for the plugin
code which is a much more flexible approach.