Use single fixture for many simulation setups
Rather than having separate fixtures for each type of experiment, I have
created a single fixture that uses parameters to enable many different
types of simulations and testing of the behaviour of the force.

New tests include ensuring that when the positions initially set away
from the expected, they will return to the expected values.